pyproject.toml
[tool.poetry]
name = "crack-the-code-game"
version = "1.0.0"
description = ""
authors = ["Vadim Tokarev <16782067+topor-dev@users.noreply.github.com>"]
license = "MIT"
repository= "https://github.com/topor-dev/crack-the-code-game"

[tool.poetry.dependencies]
python = "^3.6"
flask = "*"
attrs = "^18.2"
redis = "^3.2"

[tool.poetry.dev-dependencies]
mypy = "*"
black = "^18.9b0"

[tool.black]
skip-string-normalization = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
